Fixing a bug in SimpleCarousel: making the footer carousel dyncamically change between top and articles
